Refactor - break App to components
